Completes contact me form
