+++ portfolio/src/main/webapp/index.html
@@ -6,7 +6,6 @@
     <title>My Portfolio</title>
     <link rel="stylesheet" href="style.css">
     <script src="script.js"></script>
-    <script src="newScript.js"></script>
 </head>
 
 <body>
@@ -28,13 +27,26 @@
         <p style="text-align:center">The downtown courthouse in my home town denton</p>
         <img src="image/plants.jpg" />
         <p style="text-align:center">Some asthetically pleasing Plants</p>
-        <p>Click this button to display a fun Tv Quote:</p>
-        <button onclick="getString()">TV Quote</button>
+        <p>Click this button to display a fun Sitcom Quote:</p>
+        <button onclick="getString()">Sitcom Quote</button>
         <div id="hello-container"></div>
         <p>Click here to get a random community Quote:</p>
         <button onclick="addRandomGreeting()">Quote</button>
         <div id="greeting-container"></div>
-        <h3>Other Links.</h3>
+        <h3>Contact Me:</h3>
+        <form action="/form-handler" method="POST">
+            <p>Name:</p>
+            <textarea name="text-input"></textarea>
+            <br />
+            <p>Email:</p>
+            <textarea name="text-input2"></textarea>
+            <br />
+            <p>Message:</p>
+            <textarea name="text-input3"></textarea>
+            <br />
+            <input type="submit" />
+        </form>
+        <h4>Other Links.</h4>
         <a href="https://www.linkedin.com/in/peter-emmanuel-abah-6b02a07a/"><img src="image/linkedin.png"
                 style="width:150px;height:150px;"></a>
         <a href="https://github.com/eoa0026"><img src="image/github.png" style="width:150px;height:90px;"></a>
