+++ portfolio/src/main/webapp/style.css
@@ -28,6 +28,12 @@ h3 {
   text-align: center;
   font-size: 20px;
 }
+h4 {
+  font-family: verdana;
+  color: gray;
+  text-align: center;
+  font-size: 20px;
+}
 
 p {
   font-family: verdana;
